Refactord commands
